Fix for the undefined variable error
When completion is initialises for the very first time,
`b:completion_tried` variable doesn't appear to be defined. This
results in the following error message:

`E121 Undefined variable: b:completion_tried`

This commit fixes that error by checking if the
`b:completion_tried` variable is defined or not, before looking at its
value.