Fix an issue with omni completion not activating
Since we made the vcm_omni_pattern allowed to be set locally and
globally in commit 46a8025d4, if the local value is not set, use the
global value instead.

This fixes an issue where if the user hasn't set the b:vcm_omni_pattern
for a language, Vim wouldn't show omni completions when the user hits
Tab on things matched by `omni_pattern`.