Ensure we don't fallback until we have tried omni completion
This checks the `b:completion_tried` variable if it's set at the same
location as the previous completion. If true, then fallback to keyword
completion. Otherwise, try an omni completion instead, and then set that
we have tried the omni completion so we know to fallback for the next
time.

Resolves #40