Fix direction when omnifunc is set and vcm_direction == "p"
This commit forces the use of <C-p> <C-p> if the vcm_direction is set to
"p" and we're in omni completion mode. If this is not done, the
completion selects the first item in the list, and then subsequent Tab
presses would go back out of the popup menu and then back up the popup
menu list. Rather than that, if the vcm_direction is set to "p", the
direction is consistent, and we start at the bottom of the list, just
like keyword completion

References #51, #54